Updates to content and functionality
Rounded corners of the header, added a nav bar and made the whole header sticky.

Added more photos and implemented JS to allow users to click and expand photos, as well as click again to close the expanded view.

Embeded a spotify player for quick play.

Added social media links.

Updated the press blurbs to make the buttons more noticable.

Added 2 more music videos and adjusted the visual rules so that they stack when the screen gets smaller.